backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bidict==0.23.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-socketio==0.0.10
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic-core==2.27.2
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.45.3
typing-extensions==4.12.2
uvicorn==0.34.0
wsproto==1.2.0
